Store number of steps and number of ones in `u32` instead of `u64`
From Wikipedia we can see that for 4 states, the maximum number of
steps/ones is 107/13 respectively. For 5 states humanity has not found
the solution yet, but it's hypothesized to be 47_176_870/4098
respectively. Those numbers still fit into `u32` comfortably. And let's
be honest: no one is running this with `-n 5` to find any meaningful
answer. It would take forever.

And well, this change shrinks `Outcome` considerably, making everything
a bit faster.